+++ bin/bench-prover/src/bench_functions.rs
@@ -27,6 +27,7 @@ pub fn setup_consume_note_with_new_account() -> Result<ExecutedTransaction> {
             &[fungible_asset],
             NoteType::Public,
             None,
+            None,
         )
         .unwrap();
 
@@ -73,6 +74,7 @@ pub fn setup_consume_multiple_notes() -> Result<ExecutedTransaction> {
             &[fungible_asset_1],
             NoteType::Private,
             None,
+            None,
         )
         .unwrap();
     let note_2 = mock_chain
@@ -81,6 +83,7 @@ pub fn setup_consume_multiple_notes() -> Result<ExecutedTransaction> {
             account.id(),
             &[fungible_asset_2],
             NoteType::Private,
+            None,
             None,
         )
         .unwrap();