+++ crates/miden-testing/src/kernel_tests/batch/proposed_batch.rs
@@ -154,8 +154,14 @@ fn duplicate_unauthenticated_input_notes() -> anyhow::Result<()> {
 #[test]
 fn duplicate_authenticated_input_notes() -> anyhow::Result<()> {
     let TestSetup { mut chain, account1, account2 } = setup_chain();
-    let note =
-        chain.add_pending_p2id_note(account1.id(), account2.id(), &[], NoteType::Private, None)?;
+    let note = chain.add_pending_p2id_note(
+        account1.id(),
+        account2.id(),
+        &[],
+        NoteType::Private,
+        None,
+        None,
+    )?;
     let block1 = chain.block_header(1);
     let block2 = chain.prove_next_block();
 
@@ -195,8 +201,14 @@ fn duplicate_authenticated_input_notes() -> anyhow::Result<()> {
 #[test]
 fn duplicate_mixed_input_notes() -> anyhow::Result<()> {
     let TestSetup { mut chain, account1, account2 } = setup_chain();
-    let note =
-        chain.add_pending_p2id_note(account1.id(), account2.id(), &[], NoteType::Private, None)?;
+    let note = chain.add_pending_p2id_note(
+        account1.id(),
+        account2.id(),
+        &[],
+        NoteType::Private,
+        None,
+        None,
+    )?;
     let block1 = chain.block_header(1);
     let block2 = chain.prove_next_block();
 
@@ -274,10 +286,22 @@ fn duplicate_output_notes() -> anyhow::Result<()> {
 #[test]
 fn unauthenticated_note_converted_to_authenticated() -> anyhow::Result<()> {
     let TestSetup { mut chain, account1, account2 } = setup_chain();
-    let note0 =
-        chain.add_pending_p2id_note(account2.id(), account1.id(), &[], NoteType::Private, None)?;
-    let note1 =
-        chain.add_pending_p2id_note(account1.id(), account2.id(), &[], NoteType::Private, None)?;
+    let note0 = chain.add_pending_p2id_note(
+        account2.id(),
+        account1.id(),
+        &[],
+        NoteType::Private,
+        None,
+        None,
+    )?;
+    let note1 = chain.add_pending_p2id_note(
+        account1.id(),
+        account2.id(),
+        &[],
+        NoteType::Private,
+        None,
+        None,
+    )?;
     // The just created note will be provable against block2.
     let block2 = chain.prove_next_block();
     let block3 = chain.prove_next_block();
@@ -385,8 +409,14 @@ fn unauthenticated_note_converted_to_authenticated() -> anyhow::Result<()> {
 #[test]
 fn authenticated_note_created_in_same_batch() -> anyhow::Result<()> {
     let TestSetup { mut chain, account1, account2 } = setup_chain();
-    let note =
-        chain.add_pending_p2id_note(account1.id(), account2.id(), &[], NoteType::Private, None)?;
+    let note = chain.add_pending_p2id_note(
+        account1.id(),
+        account2.id(),
+        &[],
+        NoteType::Private,
+        None,
+        None,
+    )?;
     let block1 = chain.block_header(1);
     let block2 = chain.prove_next_block();