+++ crates/miden-testing/tests/integration/scripts/p2id.rs
@@ -45,6 +45,7 @@ fn p2id_script_multiple_assets() {
             &[fungible_asset_1, fungible_asset_2],
             NoteType::Public,
             None,
+            None,
         )
         .unwrap();
 
@@ -110,6 +111,7 @@ fn prove_consume_note_with_new_account() {
             &[fungible_asset],
             NoteType::Public,
             None,
+            None,
         )
         .unwrap();
 
@@ -158,6 +160,7 @@ fn prove_consume_multiple_notes() {
             &[fungible_asset_1],
             NoteType::Private,
             None,
+            None,
         )
         .unwrap();
     let note_2 = mock_chain
@@ -166,6 +169,7 @@ fn prove_consume_multiple_notes() {
             account.id(),
             &[fungible_asset_2],
             NoteType::Private,
+            None,
             None,
         )
         .unwrap();
@@ -208,6 +212,7 @@ fn test_create_consume_multiple_notes() {
             &[input_note_asset_1],
             NoteType::Private,
             None,
+            None,
         )
         .unwrap();
 
@@ -217,6 +222,7 @@ fn test_create_consume_multiple_notes() {
             account.id(),
             &[input_note_asset_2],
             NoteType::Private,
+            None,
             None,
         )
         .unwrap();